refactor(digest): extract shared digest-builder types/helper to a neutral module
#789: DigestBuildMeta, DigestByteLimits, and capNotes lived in plan-digest.ts,
so its sibling builders apply-digest.ts and state-digest.ts imported them
"backwards" from plan-digest -- making plan the de-facto common ancestor of two
peers. Extract the three shared pieces into a new producer-only
src/results/digest-common.ts and point every importer (both sibling builders,
base-terraform-command-handler.ts, and the five Tests/results fixtures/suites)
at it. plan-digest.ts now imports them like its siblings and no longer defines
or re-exports them.

digest-common.ts is producer-side only (the tab consumes an already-built
digest via the parity-gated digest-schema.ts / caps.ts contract), so it is
intentionally NOT added to the src/results <-> src/tab byte-identical parity
family. No behavior change: the plan/apply/state digest golden fixtures still
reproduce byte-for-byte and digest-common.js is at 100% coverage.

Closes #789
