+++ Tasks/TerraformTask/TerraformTaskV5/src/base-terraform-command-handler.ts
@@ -3,7 +3,8 @@ import { ToolRunner, IExecOptions } from 'azure-pipelines-task-lib/toolrunner';
 import { TerraformBaseCommandInitializer, TerraformAuthorizationCommandInitializer } from './terraform-commands';
 import { getSecureVarFileArgs, SecureFileLoader } from './secure-file-loader';
 import { replaceSecretFile, scrubFile, writeSecretFile } from './secure-temp';
-import { buildPlanDigest, DigestBuildMeta } from './results/plan-digest';
+import { buildPlanDigest } from './results/plan-digest';
+import { DigestBuildMeta } from './results/digest-common';
 import { buildApplyDigest, ApplyDigestOptions, parseNdjsonLines } from './results/apply-digest';
 import { buildStateDigest } from './results/state-digest';
 import { Digest } from './results/digest-schema';
